+++ src/components/Script.tsx
@@ -1,0 +1,10 @@
+export const Script = () => {
+  return (
+    <script
+      type="module"
+      dangerouslySetInnerHTML={{
+        __html: `import { hydrate } from "main";hydrate();`,
+      }}
+    />
+  );
+};