+++ src/entry.server.tsx
@@ -1,0 +1,4 @@
+import { renderToString } from "react-dom/server";
+import { type AppProps, App } from "~App";
+
+export const render = (data: AppProps) => renderToString(<App {...data} />);