+++ src/store/blog/Provider.tsx
@@ -1,9 +1,14 @@
 import { PropsWithChildren, useMemo } from "react";
+import type { BlogState } from "./state";
 import { useBlogStore } from "./store";
 import { BlogContext } from "./context";
 
-export const BlogContextProvider = (props: PropsWithChildren) => {
-  const store = useBlogStore();
+export type Props = {
+  blogs: Partial<BlogState>;
+};
+
+export const BlogContextProvider = (props: PropsWithChildren<Props>) => {
+  const store = useBlogStore(props.blogs);
   const value = useMemo(() => store, [store]);
 
   return (