+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - [#1001](./BACKLOG.md#1001) Create base react web app
 - [#1002](./BACKLOG.md#1002) Create folder based routing and components
 - [#1003](./BACKLOG.md#1003) Create script to parse, transform and generate data base on routes
+- [#1004](./BACKLOG.md#1004) Create script to build client and server js files